requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
